third_party/blink/web_tests/http/tests/devtools/profiler/heap-snapshot-summary-search-expected.txt

Tests search in Summary view of detailed heap snapshots.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testSearch

Profiler was disabled.
